created and connected the service to program cs, reorganised files struct, added leavetype controller
