+++ LeaveManagmentSystem.Web/Program.cs
@@ -1,5 +1,7 @@
 using LeaveManagmentSystem.Web.Data;
-using LeaveManagmentSystem.Web.Services;
+using LeaveManagmentSystem.Web.Services.Email;
+using LeaveManagmentSystem.Web.Services.LeaveAllocation;
+using LeaveManagmentSystem.Web.Services.LeaveTypes;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using System.Reflection;
@@ -12,6 +14,7 @@ builder.Services.AddDbContext<ApplicationDbContext>(options => options.UseSqlSer
 builder.Services.AddDatabaseDeveloperPageExceptionFilter();
 
 builder.Services.AddScoped<ILeaveTypeServices, LeaveTypeServices>();
+builder.Services.AddScoped<ILeaveAllocationService, LeaveAllocationService>();
 builder.Services.AddTransient<IEmailSender, EmailSender>();
 builder.Services.AddAutoMapper(Assembly.GetExecutingAssembly());
