+++ LeaveManagmentSystem.Web/Services/LeaveAllocation/ILeaveAllocationService.cs
@@ -1,0 +1,4 @@
+﻿namespace LeaveManagmentSystem.Web.Services.LeaveAllocation
+{
+    public interface ILeaveAllocationService { }
+}
