+++ LeaveManagmentSystem.Web/Services/LeaveAllocation/LeaveAllocationService.cs
@@ -1,0 +1,7 @@
+﻿namespace LeaveManagmentSystem.Web.Services.LeaveAllocation
+{
+    public class LeaveAllocationService(ApplicationDbContext _context) : ILeaveAllocationService
+    {
+
+    }
+}
